+++ Beam/resources/assets/js/app.js
@@ -4,6 +4,7 @@ require("remp/js/functions.js");
 require("remp/js/actions.js");
 require("remp/js/datatables.js");
 require("remp/js/vuefilters.js");
+require("remp/js/search.js");
 
 require("./forms/segmentForm.js");
 require("./forms/entityForm.js");