+++ Campaign/app/Console/Kernel.php
@@ -19,7 +19,7 @@ class Kernel extends ConsoleKernel
      */
     protected $commands = [
         Commands\CampaignsRefreshCache::class,
-        Commands\AggregateCampaignStats::class
+        Commands\AggregateCampaignStats::class,
     ];
 
     /**