+++ Campaign/resources/assets/js/app.js
@@ -5,6 +5,7 @@ require("remp/js/actions.js");
 require("remp/js/datatables.js");
 require("remp/js/chart.js");
 require("remp/js/vuefilters.js");
+require("remp/js/search.js");
 
 require("./banner.js");
 require("./forms/bannerForm.js");